+++ src/App.jsx
@@ -1,14 +1,14 @@
 import { Routes, Route } from "react-router-dom";
-import Background from "./Components/background";
-import Navbar from "./Components/navbar";
-import Header from "./Components/header";
-import Description from "./Components/description";
-import Cards from "./Components/cards";
-import CommentCards from "./Components/CommentCards";
-import UserInput from "./Components/user";
-import Footer from "./Components/footer";
-import CardHistory from "./Components/cardHistory";
-import ScrollToTopButton from "./Components/scrollToTopButton";
+import Background from "./Components/Global/background";
+import Navbar from "./Components/Global/navbar";
+import Header from "./Components/Global/header";
+import Description from "./Components/Cards/description";
+import Cards from "./Components/Cards/cards";
+import CommentCards from "./Components/Cards/commentCards";
+import UserInput from "./Components/Cards/user";
+import Footer from "./Components/Global/footer";
+import CardHistory from "./Components/History/cardHistory";
+import ScrollToTopButton from "./Components/Global/scrollToTopButton";
 import { useState } from "react";
 import "./App.css";
