+++ src/Components/Cards/cards.jsx
@@ -1,11 +1,11 @@
 import { useEffect, useState, useRef } from "react";
-import TarotCard from "../assets/images/Carta_Tarot.jpg";
-import { getCards } from "../services/ApiCards";
+import TarotCard from "../../assets/images/Carta_Tarot.jpg";
+import { getCards } from "../../services/ApiCards";
 import SelectedCards from "./selectedCards";
-import CardsResult from "./CardsResult";
-import AlertPopup from "./modals/popupAlert";
-import StartModal from "./modals/startModal";
-import EndModal from "./modals/endModal";
+import CardsResult from "./cardsResult";
+import AlertPopup from "../modals/popupAlert";
+import StartModal from "../modals/startModal";
+import EndModal from "../modals/endModal";
 
 const shuffleArray = (array) => {
   const shuffled = [...array];
