+++ src/Components/Global/scrollToTopButton.jsx
@@ -28,7 +28,7 @@ export default function ScrollToTopButton() {
     <button
       onClick={scrollToTop}
       className={`fixed bottom-6 right-6 z-50 transition-opacity duration-300 
-        ${isVisible ? "opacity-100" : "opacity-0 pointer-events-none"}`}
+        ${isVisible ? "opacity-100" : "opacity-0 pointer-events-none"} cursor-pointer`}
     >
       <FaCircleArrowUp
         className="w-12 h-12 text-white drop-shadow-lg hover:text-[#C99AE9] active:scale-95 transition-transform"
