+++ src/Components/History/clearHistory.jsx
