+++ src/services/ApiHistory.js
@@ -20,7 +20,6 @@ export const updateUserName = async (id, newUserName) => {
   return await axios.patch(`${API_URL}/${id}`, { user: newUserName });
 };
 
-// 👉 Borrar todas las tiradas una a una
 export const deleteAllSpreads = async () => {
   const spreads = await getSpreads();
   const deletePromises = spreads.map((s) => deleteSpread(s.id));
